Added menu item for toggling domain based exemption.
The menu item dynamically updates its label based on the current domain
and updates a permanent file in the users preference folder to store
the domain exemption list.

The window has a listener that updates known exemptions whenever focus
is reacquired.  This is to handle cases where the exemption list is
modified outside of the current window either through the toggle feature
in another window or possibly direct text editing of the config file.

Most helper functions and objects are store in window.dimmerAddon.
This may or may not be better than storing the required functions in
the addon's global scope.  Further investigation is required.  Once a
clear path is chosen the code can be updated to do things in a
consistent way.