Fixed unloading issues and added dynamic redraws
The preference observer wasn't unloading.  Referencing it as an element
of the window object rather than on the global scope fixed the problem.

When preferences are changed the currently open pages will have their
dimmer overlays redrawn to the new opacity.

When exemption is toggled for a domain, the currently open page will
have its overlay updated.  This should actually updated all currently
open windows and tabs because pages from the same domain could be open
in different tabs.